requirements.txt
openai
transformers
pdfminer.six
